added compiler flags for
- warnings on unused imports
- warnings on numeric widening
- disabling adapted args (automatic tuple conversion)
- fail the build on compiler warnings

Overall this should lead to increased code quality.

+++ src/main/scala/mesosphere/marathon/MarathonSchedulerDriver.scala
@@ -22,7 +22,7 @@ object MarathonSchedulerDriver {
                 frameworkId: Option[FrameworkID]): SchedulerDriver = {
     val frameworkInfoBuilder = FrameworkInfo.newBuilder()
       .setName(frameworkName)
-      .setFailoverTimeout(config.mesosFailoverTimeout())
+      .setFailoverTimeout(config.mesosFailoverTimeout().toDouble)
       .setUser(config.mesosUser())
       .setCheckpoint(config.checkpoint())
